requirements.txt
streamlit
langchain
PyPDF2
faiss-cpu
cohere
python-dotenv
pyprojroot